Restructure of source tree
This is a major restructuring of the source tree for the following goals:

1. Promote code compartmentalization
2. Define a clear interface between scenic business logic and rendering logic
3. Improve debug logging
4. Improve long term code maintainability
5. Facilitate adding new devices

Much of the code is simply moved around as part of the reorganization.
Weak functions are used in the script_ops.c to provide a default function
that simply logs that the function was called. This is to help with writing
new device drivers, were the implementer can focus on those functions that
matter to them first.

This is probably best reviewed by examining the whole source tree rather
than the diff.

I have tested this with glfw and seems to work fine